requirements.txt
fastapi
uvicorn
python-dotenv
supabase
fuzzywuzzy
pdf2image
Pillow
PyPDF2
python-multipart
requests